+++ tests/test_md.py
@@ -21,7 +21,7 @@ def test_virial_tensor():
     # Normally, the parser should enforce just returning the names of atom types
     atom_indices_to_names = {0: "O_water", 1: "H_water"}
     atom_type_names = [atom_indices_to_names[int(atom_types[i])] for i in range(len(atom_types))]
-    box = torch.tensor(np.eye(3) * 18.643 / BOHR2ANG, dtype=torch.float64, requires_grad=True)
+    box = torch.tensor(np.eye(3) * 18.643 / BOHR2ANG, requires_grad=True)
     box_volume = torch.det(box)
 
     cm = CoordinateManager(coords, box, 10.0 / BOHR2ANG, 1024)
@@ -52,7 +52,7 @@ def test_md():
     # Normally, the parser should enforce just returning the names of atom types
     atom_indices_to_names = {0: "O_water", 1: "H_water"}
     atom_type_names = [atom_indices_to_names[int(atom_types[i])] for i in range(len(atom_types))]
-    box = torch.tensor(np.eye(3) * 18.643 / BOHR2ANG, dtype=torch.float64, requires_grad=True)
+    box = torch.tensor(np.eye(3) * 18.643 / BOHR2ANG, requires_grad=True)
     cm = CoordinateManager(coords, box, 10.0 / BOHR2ANG, 1024)
     topology = Topology(bonds, cm.neighbor_list, coords.size(0))
     pairs, dists, dist_vecs = cm.get_distances_vectors_and_pairs()
@@ -73,7 +73,7 @@ def test_optimize_nacl():
 
     positions = np.loadtxt(os.path.join(os.path.dirname(__file__), "data/nacl_crystal.txt"), dtype=np.float64)
     positions[0, 0] += 0.1 # move from equilibrium
-    coords = torch.tensor(positions / BOHR2NM, dtype=torch.float64, requires_grad=True)
+    coords = torch.tensor(positions / BOHR2NM, requires_grad=True)
     bonds = np.array([], dtype=np.float64)
     atom_type_names = ["" for i in range(coords.size(0))]
     for i in range(coords.size(0) // 2):
@@ -81,7 +81,7 @@ def test_optimize_nacl():
     for i in range(coords.size(0) // 2, coords.size(0)):
         atom_type_names[i] = "Cl-"
 
-    box = torch.tensor(np.eye(3) * 28.2 / BOHR2ANG, dtype=torch.float64, requires_grad=True)
+    box = torch.tensor(np.eye(3) * 28.2 / BOHR2ANG, requires_grad=True)
     cm = CoordinateManager(coords, box, 10.0 / BOHR2ANG, 2048)
     pairs, dists, dist_vecs = cm.get_distances_vectors_and_pairs()
     with torch.no_grad():
